# Enable hotplugging joysticks