--- a/include/mapper.h
+++ b/include/mapper.h
@@ -87,4 +87,17 @@ void MAPPER_CheckEvent(SDL_Event *event);
 // for good measure.
 constexpr int MaxBindNameLength = 100;
 
+/**
+ * Handle a game controller being connected or disconnected by reiniitializing
+ * binds. Caller is responsible for any user interface updates, e.g. updating
+ * the mapper UI buttons to reflect joystick binds being added / removed.
+ *
+ * @param event A pointer to an SDL_JoyDeviceEvent whose type must be one of:
+ * - SDL_CONTROLLERDEVICEADDED
+ * - SDL_CONTROLLERDEVICEREMOVED
+ * - SDL_JOYDEVICEADDED
+ * - SDL_JOYDEVICEREMOVED
+ */
+void MAPPER_HandleJoyDeviceEvent(SDL_JoyDeviceEvent* event);
+
 #endif