--- a/include/video.h
+++ b/include/video.h
@@ -69,8 +69,6 @@
 //   postfixes is highly recommended in such cases to remove ambiguity.
 //
 
-#define REDUCE_JOYSTICK_POLLING
-
 enum class RenderingBackend {
 	Texture,
 	OpenGl
@@ -432,9 +430,7 @@ void GFX_SetMouseRawInput(const bool requested_raw_input);
 // Detects the presence of a desktop environment or window manager
 bool GFX_HaveDesktopEnvironment();
 
-#if defined(REDUCE_JOYSTICK_POLLING)
 void MAPPER_UpdateJoysticks(void);
-#endif
 
 DosBox::Rect GFX_CalcDrawRectInPixels(const DosBox::Rect& canvas_size_px,
                                       const DosBox::Rect& render_size_px,