--- a/src/gui/sdl_mapper.cpp
+++ b/src/gui/sdl_mapper.cpp
@@ -732,10 +732,8 @@ public:
 
 			if (event->jaxis.which != stick_id)
 				return nullptr;
-#if defined(REDUCE_JOYSTICK_POLLING)
 			if (axis_id >= axes)
 				return nullptr;
-#endif
 			if (abs(axis_position) < 25000)
 				return nullptr;
 
@@ -747,11 +745,7 @@ public:
 		} else if (event->type == SDL_JOYBUTTONDOWN) {
 			if (event->jbutton.which != stick_id)
 				return nullptr;
-#if defined (REDUCE_JOYSTICK_POLLING)
-			return CreateButtonBind(event->jbutton.button%button_wrap);
-#else
-			return CreateButtonBind(event->jbutton.button);
-#endif
+			return CreateButtonBind(event->jbutton.button % button_wrap);
 		} else if (event->type==SDL_JOYHATMOTION) {
 			if (event->jhat.which != stick_id) return nullptr;
 			if (event->jhat.value == 0) return nullptr;
@@ -777,15 +771,16 @@ public:
 			case SDL_JOYBUTTONDOWN:
 			case SDL_JOYBUTTONUP:
 				jbutton = &event->jbutton;
-				bool state;
-				state = jbutton->type == SDL_JOYBUTTONDOWN;
-				const auto but = check_cast<uint8_t>(jbutton->button % emulated_buttons);
-				if (jbutton->which == stick_id) {
-					JOYSTICK_Button(emustick, but, state);
-				}
-				break;
-			}
-			return false;
+			        if (jbutton->which != stick_id) {
+				        break;
+			        }
+			        const bool state = jbutton->type == SDL_JOYBUTTONDOWN;
+			        const auto but = check_cast<uint8_t>(
+			                jbutton->button % emulated_buttons);
+			        JOYSTICK_Button(emustick, but, state);
+			        break;
+		        }
+		        return false;
 	}
 
 	virtual void UpdateJoystick() {
@@ -2637,8 +2632,8 @@ static void ClearAllBinds()
 	}
 }
 
-static void CreateDefaultBinds() {
-	ClearAllBinds();
+static void CreateDefaultBinds()
+{
 	char buffer[512];
 	Bitu i=0;
 	while (DefaultKeys[i].eventend) {
@@ -2760,7 +2755,6 @@ static bool load_binds_from_file(const std::string_view mapperfile_path,
 
 		suppress_save_mapper_file_message = true;
 
-		ClearAllBinds();
 		for (auto &line : lines)
 			CreateStringBind(line.data());
 
@@ -2787,17 +2781,68 @@ static bool load_binds_from_file(const std::string_view mapperfile_path,
 
 void MAPPER_CheckEvent(SDL_Event *event)
 {
-	for (auto &group : bindgroups)
-		if (group->CheckEvent(event))
+	switch (event->type) {
+	case SDL_CONTROLLERDEVICEADDED:
+	case SDL_CONTROLLERDEVICEREMOVED:
+	case SDL_JOYDEVICEREMOVED:
+	case SDL_JOYDEVICEADDED:
+		MAPPER_HandleJoyDeviceEvent(
+		        reinterpret_cast<SDL_JoyDeviceEvent*>(event));
+		return;
+	default: break;
+	}
+
+	MAPPER_UpdateJoysticks();
+	for (auto& group : bindgroups) {
+		if (group->CheckEvent(event)) {
 			return;
+		}
+	}
 }
 
-void BIND_MappingEvents() {
+void MAPPER_HandleJoyDeviceEvent(SDL_JoyDeviceEvent* event)
+{
+	switch (event->type) {
+	case SDL_CONTROLLERDEVICEADDED:
+	case SDL_JOYDEVICEADDED: {
+		auto index    = event->which;
+		auto joystick = SDL_JoystickOpen(index);
+		auto name     = SDL_JoystickName(joystick);
+		LOG_INFO("MAPPER: Game controller '%s' has been plugged in", name);
+		SDL_JoystickClose(joystick);
+		break;
+	}
+	case SDL_CONTROLLERDEVICEREMOVED:
+	case SDL_JOYDEVICEREMOVED: {
+		auto joystick = SDL_JoystickFromInstanceID(event->which);
+		auto name     = SDL_JoystickName(joystick);
+		LOG_INFO("MAPPER: Game controller '%s' has been disconnected", name);
+		SDL_JoystickClose(joystick);
+		break;
+	}
+	default:
+		LOG_DEBUG("MAPPER:  MAPPER_HandleJoyDeviceEvent: Unexpected event type '%d' received",
+		          event->type);
+		assert(false);
+	}
+	MAPPER_BindKeys(get_sdl_section());
+}
+
+void BIND_MappingEvents()
+{
 	SDL_Event event;
 	static bool isButtonPressed = false;
 	static CButton *lastHoveredButton = nullptr;
 	while (SDL_PollEvent(&event)) {
 		switch (event.type) {
+		case SDL_CONTROLLERDEVICEADDED:
+		case SDL_CONTROLLERDEVICEREMOVED:
+		case SDL_JOYDEVICEREMOVED:
+		case SDL_JOYDEVICEADDED:
+			MAPPER_HandleJoyDeviceEvent(
+			        reinterpret_cast<SDL_JoyDeviceEvent*>(&event));
+			mapper.redraw = true;
+			break;
 		case SDL_MOUSEBUTTONDOWN:
 			isButtonPressed = true;
 			/* Further check where are we pointing at right now */
@@ -2884,6 +2929,17 @@ void BIND_MappingEvents() {
 	}
 }
 
+static void ClearSticks()
+{
+	// Free any allocated sticks
+	for (int i = 0; i < MAXSTICKS; ++i) {
+		delete mapper.sticks.stick[i];
+		mapper.sticks.stick[i] = nullptr;
+	}
+	mapper.sticks.num        = 0;
+	mapper.sticks.num_groups = 0;
+}
+
 //  Initializes SDL's joystick subsystem an setups up initial joystick settings.
 
 // If the user wants auto-configuration, then this sets joytype based on queried
@@ -2896,7 +2952,7 @@ void BIND_MappingEvents() {
 static void QueryJoysticks()
 {
 	// Reset our joystick status
-	mapper.sticks.num = 0;
+	ClearSticks();
 
 	JOYSTICK_ParseConfiguredType();
 
@@ -2972,8 +3028,23 @@ static void QueryJoysticks()
 	}
 }
 
-static void CreateBindGroups() {
+static void ClearBindGroups()
+{
+	for (auto& ptr : keybindgroups) {
+		delete ptr;
+	}
+	keybindgroups.clear();
+
+	for (auto& ptr : stickbindgroups) {
+		delete ptr;
+	}
+	stickbindgroups.clear();
+
 	bindgroups.clear();
+}
+
+static void CreateBindGroups()
+{
 	CKeyBindGroup* key_bind_group = new CKeyBindGroup(SDL_NUM_SCANCODES);
 	keybindgroups.push_back(key_bind_group);
 
@@ -2983,22 +3054,6 @@ static void CreateBindGroups() {
 		return;
 
 	if (joytype != JOY_NONE_FOUND) {
-#if defined (REDUCE_JOYSTICK_POLLING)
-		// direct access to the SDL joystick, thus removed from the event handling
-		if (mapper.sticks.num)
-			SDL_JoystickEventState(SDL_DISABLE);
-#else
-		// enable joystick event handling
-		if (mapper.sticks.num)
-			SDL_JoystickEventState(SDL_ENABLE);
-		else
-			return;
-#endif
-		// Free up our previously assigned joystick slot before assinging below
-		if (mapper.sticks.stick[mapper.sticks.num_groups]) {
-			delete mapper.sticks.stick[mapper.sticks.num_groups];
-			mapper.sticks.stick[mapper.sticks.num_groups] = nullptr;
-		}
 
 		uint8_t joyno = 0;
 		switch (joytype) {
@@ -3047,18 +3102,12 @@ static void CreateBindGroups() {
 	}
 }
 
-bool MAPPER_IsUsingJoysticks() {
-	return (mapper.sticks.num > 0);
-}
-
-#if defined (REDUCE_JOYSTICK_POLLING)
 void MAPPER_UpdateJoysticks() {
 	for (Bitu i=0; i<mapper.sticks.num_groups; i++) {
 		assert(mapper.sticks.stick[i]);
 		mapper.sticks.stick[i]->UpdateJoystick();
 	}
 }
-#endif
 
 void MAPPER_LosingFocus() {
 	for (const auto& event : events) {
@@ -3179,9 +3228,6 @@ void MAPPER_DisplayUI() {
 	mapper.exit = false;
 	mapper.redraw=true;
 	SetActiveEvent(nullptr);
-#if defined (REDUCE_JOYSTICK_POLLING)
-	SDL_JoystickEventState(SDL_ENABLE);
-#endif
 	while (!mapper.exit) {
 		if (mapper.redraw) {
 			mapper.redraw = false;
@@ -3214,9 +3260,6 @@ void MAPPER_DisplayUI() {
 		}
 	}
 #endif
-#if defined (REDUCE_JOYSTICK_POLLING)
-	SDL_JoystickEventState(SDL_DISABLE);
-#endif
 	GFX_ResetScreen();
 	MOUSE_NotifyTakeOver(false);
 }
@@ -3235,22 +3278,10 @@ static void MAPPER_Destroy(Section *sec) {
 
 	buttons.clear();
 
-	for (auto & ptr : keybindgroups)
-		delete ptr;
-	keybindgroups.clear();
-
-	for (auto & ptr : stickbindgroups)
-		delete ptr;
-	stickbindgroups.clear();
-
-	// Free any allocated sticks
-	for (int i = 0; i < MAXSTICKS; ++i) {
-		delete mapper.sticks.stick[i];
-		mapper.sticks.stick[i] = nullptr;
-	}
+	ClearBindGroups();
+	ClearSticks();
 
 	// Empty the remaining lists now that their pointers are defunct
-	bindgroups.clear();
 	handlergroup.clear();
 	holdlist.clear();
 
@@ -3292,14 +3323,15 @@ void MAPPER_BindKeys(Section* sec)
 	assert(property);
 	mapper.filename = property->realpath.string();
 
+	ClearAllBinds();
+	ClearBindGroups();
 	QueryJoysticks();
 
 	// Create the graphical layout for all registered key-binds
 	if (buttons.empty())
 		CreateLayout();
 
-	if (bindgroups.empty())
-		CreateBindGroups();
+	CreateBindGroups();
 
 	// Create binds from file or fallback to internals
 	if (!load_binds_from_file(mapper.filename, mapperfile_value))
@@ -3314,8 +3346,6 @@ void MAPPER_BindKeys(Section* sec)
 
 	if (SDL_GetModState()&KMOD_NUM)
 		MAPPER_TriggerEvent(num_lock_event, false);
-
-	GFX_RegenerateWindow(sec);
 }
 
 std::vector<std::string> MAPPER_GetEventNames(const std::string &prefix) {
@@ -3341,6 +3371,7 @@ void MAPPER_StartUp(Section* sec)
 	// mapperfile=file.map"` commands
 	constexpr auto changeable_at_runtime = true;
 	section->AddInitFunction(&MAPPER_BindKeys, changeable_at_runtime);
+	section->AddInitFunction(&GFX_RegenerateWindow, changeable_at_runtime);
 
 	// Runs one-time on shutdown
 	section->AddDestroyFunction(&MAPPER_Destroy);