--- a/src/misc/setup.cpp
+++ b/src/misc/setup.cpp
@@ -1980,3 +1980,24 @@ bool config_file_is_valid(const std_fs::path& path)
 	fclose(file);
 	return false;
 }
+
+// Get up-to-date in-memory model of a config section
+static Section_prop* get_section(const char* section_name)
+{
+	assert(control);
+
+	const auto sec = static_cast<Section_prop*>(control->GetSection(section_name));
+	assert(sec);
+
+	return sec;
+}
+
+Section_prop* get_joystick_section()
+{
+	return get_section("joystick");
+}
+
+Section_prop* get_sdl_section()
+{
+	return get_section("sdl");
+}