# Description

* fixes #946
* remove REDUCE_JOYSTICK_POLLING hack, which prevented detection of joystick hotplug events
* re-initialize mapper when receiving following events:
  * SDL_CONTROLLERDEVICEADDED
  * SDL_CONTROLLERDEVICEREMOVED
  * SDL_JOYDEVICEADDED
  * SDL_JOYDEVICEREMOVED

## Related issues

* #946
* #4124


# Release notes

* Add option to enable hotplugging of joystick devices after program start (enabled by default)

# Manual testing

The change has been manually tested on:

- [x] Windows
- [x] macOS
- [X] Linux

I tested on Ubuntu 24.04 with XBox 360 Wireless Controller (ID 045e:0719 Microsoft Corp. Xbox 360 Wireless Adapter).
I ran DOSBox with and without controller turned on.  I verified that controller still functioned in game in each case after power cycling the controller (which causes device to disconnect and reconnect).

Built using meson on Clion.

Subsequently, I tested on Windows using VS 2022 community edition.

Afterwards, I tested on macOS Ventura 13.7.1. I built using Homebrew. Notably, I needed to `brew install glib`, which was not listed in the list of brew dependencies in the instructions. On macOS the xbox 360 wireless receiver does not work, so instead I tested with:
```
Xbox Wireless Controller:
  Vendor ID:	0x045E
  Product ID:	0x0B13
```

# Checklist

I have:

- [x] followed the project's [contributing guidelines](https://github.com/dosbox-staging/dosbox-staging/blob/master/CONTRIBUTING.md) and [code of conduct](https://github.com/dosbox-staging/dosbox-staging/blob/master/CODE_OF_CONDUCT.md).
- [x] performed a self-review of my code.
- [x] commented on the particularly hard-to-understand areas of my code.
- [x] split my work into well-defined, bisectable commits, and I [named my commits well](https://github.com/dosbox-staging/dosbox-staging/blob/main/CONTRIBUTING.md#commit-messages).
- [x] applied the appropriate labels (bug, enhancement, refactoring, documentation, etc.)
- [x] [checked](https://github.com/dosbox-staging/dosbox-staging/blob/main/scripts/compile_commits.sh) that all my commits can be built.
- [x] my change has been manually tested on Windows, macOS, and Linux.
- [x] confirmed that my code does not cause performance regressions (e.g., by running the Quake benchmark).
- [ ] added unit tests where applicable to prove the correctness of my code and to avoid future regressions.
- [x] provided the release notes draft (for significant user-facing changes).
- [ ] made corresponding changes to the documentation or the website according to the [documentation guidelines](https://github.com/dosbox-staging/dosbox-staging/blob/main/DOCUMENTATION.md).
- [ ] [locally verified](https://github.com/dosbox-staging/dosbox-staging/blob/main/DOCUMENTATION.md#previewing-documentation-changes-locally) my website or documentation changes.