@johnnovak 2024-12-27
Nice enhancement @arifogel, we'll test it on all 3 OSes with the team 🚀 

The code looks good, mostly just a few questions about the joystick/controller distinction (probably due to my unfamiliarity with the input handling in SDL).

Thanks for filling out the PR template correctly 💯  Could you please remove the placeholder texts from the start of the sections?