@FeralChild64 2024-12-27
_Linux_ - I confirm it works; whether I start DOSBox with disconnected pad and connect it later, or whether I start DOSBox with a connected pad, disconnect it and reconnect it again, I can use the pad in Wolfenstein 3D.

_Windows_ - DOSBox is unusable with this patch, it constantly re-initializes the display. It looks like it thinks the pad gets constantly disconnected and reconnected again:

 ```
2024-12-27 16:43:51.702 | MAPPER: Game controller 'Nintendo Switch Pro Controller' has been plugged in
2024-12-27 16:43:51.822 | MAPPER: Found one joystick
2024-12-27 16:43:51.942 | MAPPER: Initialised Nintendo Switch Pro Controller with 6 axes, 20 buttons, and 0 hat(s)
2024-12-27 16:43:51.943 | MAPPER: Loaded default key bindings
2024-12-27 16:43:51.949 | DISPLAY: Using 1219x914 window size in windowed mode on display-0
2024-12-27 16:43:53.126 | OPENGL: Vendor: Intel
2024-12-27 16:43:53.126 | OPENGL: Version: 4.6.0 - Build 32.0.101.5972
2024-12-27 16:43:53.126 | OPENGL: GLSL version: 4.60 - Build 32.0.101.5972
2024-12-27 16:43:53.126 | OPENGL: NPOT textures supported
2024-12-27 16:43:53.129 | DISPLAY: VGA 720x400 16-colour text mode 03h at 70.087 Hz VFR, scaled to 2133x1600 pixels with 1:1.35 (20:27) pixel aspect ratio
2024-12-27 16:43:55.005 | MAPPER: Game controller 'Nintendo Switch Pro Controller' has been disconnected
2024-12-27 16:43:55.006 | MAPPER: No joysticks found
2024-12-27 16:43:55.006 | MAPPER: Loaded default key bindings
2024-12-27 16:43:55.008 | DISPLAY: Using 1219x914 window size in windowed mode on display-0
2024-12-27 16:43:56.172 | OPENGL: Vendor: Intel
2024-12-27 16:43:56.173 | OPENGL: Version: 4.6.0 - Build 32.0.101.5972
2024-12-27 16:43:56.173 | OPENGL: GLSL version: 4.60 - Build 32.0.101.5972
2024-12-27 16:43:56.173 | OPENGL: NPOT textures supported
2024-12-27 16:43:56.175 | DISPLAY: VGA 720x400 16-colour text mode 03h at 70.087 Hz VFR, scaled to 2133x1600 pixels with 1:1.35 (20:27) pixel aspect ratio
2024-12-27 16:43:56.294 | MAPPER: Game controller 'Nintendo Switch Pro Controller' has been plugged in
2024-12-27 16:43:56.414 | MAPPER: Found one joystick
2024-12-27 16:43:56.534 | MAPPER: Initialised Nintendo Switch Pro Controller with 6 axes, 20 buttons, and 0 hat(s)
2024-12-27 16:43:56.535 | MAPPER: Loaded default key bindings
2024-12-27 16:43:56.540 | DISPLAY: Using 1219x914 window size in windowed mode on display-0
2024-12-27 16:43:57.749 | OPENGL: Vendor: Intel
2024-12-27 16:43:57.749 | OPENGL: Version: 4.6.0 - Build 32.0.101.5972
2024-12-27 16:43:57.749 | OPENGL: GLSL version: 4.60 - Build 32.0.101.5972
2024-12-27 16:43:57.749 | OPENGL: NPOT textures supported
2024-12-27 16:43:57.752 | DISPLAY: VGA 720x400 16-colour text mode 03h at 70.087 Hz VFR, scaled to 2133x1600 pixels with 1:1.35 (20:27) pixel aspect ratio
```

I'll try _macOS_ later, my _Windows_ has just started installing some huge update, I have no idea when it ends.