@weirddan455 2024-12-27
@FeralChild64 Is it supposed to re-init the display when a controller gets plugged/unplugged?  That seems unnecessary.