@FeralChild64 2024-12-27
@weirddan455 This is definitely undesired. Especially that with botched pad/OS/chipset/drivers/whatever (I really don't want to dig into it to check which component caused the problem) the user might easily end up with non-working DOSBox.

BTW, I can confirm the patch works correctly on my macOS device.