@arifogel 2024-12-27
@FeralChild64 @weirddan455 The display re-initing was slightly annoying for me, and on at least one OS it knocked me out of full screen when those events were received. Given the problems @FeralChild64 encountered in Windows fpr constant reconnects/disconnects, I will spend a bit more time trying to improve the UX on disconnect/reconnect for this patch, i.e. preventing display refresh.

Another option perhaps in the short term is to just make controller hotplugging optional and disabled by default, if it's too much work to factor out the display reinitialization. But I will try the better solution first.