@arifogel 2024-12-27
There is also one more thing I'd like to check so I don't forget:

- [ ] Confirm binds that have been changed but not saved remain changed following controller connect/disconnect event