@arifogel 2024-12-27
Updates on latest commit::
1. Display is no longer reinitialized when joystick is connected or disconnect
1. Joystick hotplug events are no longer ignored in Mapper UI. Mapper UI is redrawn after joystick hotplug event
1. Refactored a bit to clean up and avoid code duplication