@arifogel 2024-12-27
> There is also one more thing I'd like to check so I don't forget:
> 
> * [ ]  Confirm binds that have been changed but not saved remain changed following controller connect/disconnect event

So I tested this, and it looks like unsaved binds are thrown away (even unrelated keyboard binds) on controller hotplug event. On the one hand this might be very annoying for someone who has set some binds, played a bit, then set down the controller for 5 minutes to find that the binds are gone on reconnect.

I can think of a couple solutions that are all a bit wonky:
1. Leave as is
2. Remember in-memory changes and re-apply them on top of reinitialized binds (somewhat involved, merge strategy non-trivial)
3. Just always save binds to disk whenever they are changed instead of having a save button (non-trivial but probably not difficult, but is a major opinionated UX change)

-----

> Joystick hotplug events are no longer ignored in Mapper UI. Mapper UI is redrawn after joystick hotplug event

I just realized that with this change, @FeralChild64 's constantly hotplugging controller would render the mapper UI unusable. For the use case of naughty controllers a user doesn't want to have to disconnect from the their computer to use this program, I'm thinking to just make hotplugging configurable. We can also have a keyboard shortcut or something to manually redetect controllers when hotplugging is disabled.

Thoughts?