@arifogel 2024-12-27
Anyway in latest commit I added  (and documented) a configuration option to control whether hotplugging is enabled (default to true), to provide an out for those with misbehaving controllers. Going to put this down until next round of review.