@arifogel 2024-12-28
Because this branch removes a joystick polling hack that allegedly improves performance, I decided to run the first quake benchmark from dosbench on various configurations to see the impact of these changes are performance.

I've attached my dosbox-staging.conf file.

* Each benchmark was done in full screen.
* `Joystick calibrated properly` means during benchmark, joystick was calibrated by pressing button 1 with axis in center position, then again with stick held in upper left corner, then again with stick held in bottom right corner. A `no` in the corresponding box means that button 1 was pressed 3 times while the stick was loose the whole time.
* `Joystick stick held` means that the stick was held in the bottom right corner for the duration of the benchmark after calibration. A `no` in the box means the stick was left loose following calibration until the end of the benchmark
* `PR` results are from my  [hotplug branch at this tag](https://github.com/arifogel/dosbox-staging/tree/ari-hotplug-joystick-f99bb9189)
* `main` results are from 256c06ebc

### Linux results

|Joystick present|Joystick calibrated properly|Joystick stick held|PR FPS|main FPS|PR Duration|main Duration|
|-|-|-|-|-|-|-|
|no|N/A|N/A|351.9|356.9|2.8s|2.7|
|yes|yes|yes|261.9|265.2|3.7|3.7|
|yes|yes|no|303.3|298.4|3.2|3.2|
|yes|no|no|305.5|306.5|3.2|3.2|

So it seems on Linux, while joystick use does have an impact on performance, the impact of removing the polling hack in this PR is negligible or nonexistent.

macOS results will follow, since the hack was allegedly targeted for macOS originally.