@arifogel 2024-12-28
### MacOS Benchmark Results

|Joystick present|Joystick calibrated properly|Joystick stick held|PR FPS|main FPS|PR Duration|main Duration|
|-|-|-|-|-|-|-|
|no|N/A|N/A|190.5|194.7|5.1|5.0|
|yes|yes|yes|153.1|152.8|6.3|6.3|
|yes|yes|no|168.0|172.7|5.8|5.6|
|yes|no|no|171.8|170.4|5.6|5.7|

So for macOS, we again we have an inconsistent/negligible impact of the PR on performance.

[dosbox-staging.conf.gz](https://github.com/user-attachments/files/18264959/dosbox-staging.conf.gz)